Fix three Kotlin compilation errors
- NumberChart: cast Double to Float for FloatEntry (ChartSeries uses Double)
- CreateActivityViewModel: assignments are not expressions in Kotlin;
  convert single-expression setters (= assignment) to block bodies ({ })
- SettingsViewModel: private launch() returned Job, causing KFunction1<Bool,Job>
  vs (Bool)->Unit mismatch on Switch.onCheckedChange; use block body to return Unit

+++ app/src/main/java/com/planner/app/feature/analytics/components/NumberChart.kt
@@ -23,7 +23,7 @@ fun NumberChart(
     if (series.dataPoints.isEmpty()) return
 
     val entries = remember(series.dataPoints) {
-        series.dataPoints.mapIndexed { i, (_, v) -> FloatEntry(i.toFloat(), v) }
+        series.dataPoints.mapIndexed { i, (_, v) -> FloatEntry(i.toFloat(), v.toFloat()) }
     }
     val model = remember(entries) { entryModelOf(entries) }
